LayoutModule: initial commit

+++ libs/Venne/CMS/Modules/AliasModule/AliasModel.php
@@ -84,8 +84,10 @@ class AliasModel extends Venne\CMS\Developer\Model {
 	public function removeItemByModuleName($moduleName, $moduleItemId)
 	{
 		$item = $this->getRepository()->findOneBy(array("moduleName"=>$moduleName, "moduleItemId"=>$moduleItemId));
-		$this->getEntityManager()->remove($item);
-		$this->getEntityManager()->flush();
+		if($item){
+			$this->getEntityManager()->remove($item);
+			$this->getEntityManager()->flush();
+		}
 	}
 	
 }
