+++ libs/Venne/CMS/Modules/CommentsModule/CommentsModel.php
@@ -65,8 +65,10 @@ class CommentsModel extends Venne\CMS\Developer\Model {
 	public function removeItemByModuleName($moduleName, $moduleItemId)
 	{
 		$item = $this->getRepository()->findOneBy(array("moduleName"=>$moduleName, "moduleItemId"=>$moduleItemId));
-		$this->getEntityManager()->remove($item);
-		$this->getEntityManager()->flush();
+		if($item){
+			$this->getEntityManager()->remove($item);
+			$this->getEntityManager()->flush();
+		}
 	}
 	
 }
